backend/order_service/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.28.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic[email]==2.4.2
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
httpx==0.25.1
python-multipart==0.0.6
email-validator==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.10.1
isort==5.12.0
mypy==1.6.1 